+++ src/datagen/java/blusunrize/immersiveengineering/data/recipes/ClocheRecipes.java
@@ -227,7 +227,7 @@ public class ClocheRecipes extends IERecipeProvider
 				.build(out, toRL("cloche/chorus_fruit"));
 		ClocheRecipeBuilder.builder()
 				.output(Ingredients.HEMP_FIBER)
-				.output(new ItemStack(Misc.HEMP_SEEDS, 2))
+				.output(Misc.HEMP_SEEDS, 0.5f)
 				.seed(Misc.HEMP_SEEDS)
 				.soil(Blocks.DIRT)
 				.setTime(800)
@@ -264,7 +264,7 @@ public class ClocheRecipes extends IERecipeProvider
 	{
 		ClocheRecipeBuilder.builder()
 				.output(Items.WHEAT, 2)
-				.output(Items.WHEAT_SEEDS, 1)
+				.output(Items.WHEAT_SEEDS, 0.25f)
 				.seed(Items.WHEAT_SEEDS)
 				.soil(Blocks.DIRT)
 				.setTime(640)
@@ -286,7 +286,7 @@ public class ClocheRecipes extends IERecipeProvider
 				.build(out, toRL("cloche/carrot"));
 		ClocheRecipeBuilder.builder()
 				.output(new ItemStack(Items.BEETROOT, 2))
-				.output(new ItemStack(Items.BEETROOT_SEEDS, 1))
+				.output(Items.BEETROOT_SEEDS, 0.25f)
 				.seed(Items.BEETROOT_SEEDS)
 				.soil(Blocks.DIRT)
 				.setTime(800)