+++ changelog.md
@@ -1,5 +1,6 @@
 ##### Version 1.21.1-12.3.2-pre
-- Simplify a lot of recipe JSONs
+- Add support for chance-based output to the cloche (BluSunrize) 
+- Simplify a lot of recipe JSONs (BluSunrize)
     - The wrapping in "basePredicate" is now not needed for normal items anymore, and "count" on item stacks is optional too
 
 ##### Version 1.21.1-12.3.1-189